requirements.txt
streamlit
psycopg2-binary
google-generativeai
python-dotenv
sqlalchemy